Move tests to new directories
